requirements.txt
Django==4.2.1
djangorestframework==3.14.0
django-q==1.3.9
django-extensions==3.2.1
django-filter==23.2
ipython==8.13.2
Faker==18.7.0